Refactor the export api and improve error messages (#54)
Improved error messages to provide clear next steps when there is an
export error

```
E           torch_onnx.errors.OnnxConversionError: Failed to convert the exported program to an ONNX model. This is step 2/2 of exporting the model to ONNX. Next steps:
E           - If there is a missing ONNX function, implement it and register it to the registry.
E           - If there is an internal error during ONNX conversion, debug the error and summit a PR to PyTorch.
E           - Save the ExportedProgram as a pt2 file and create an error report with `export(error_report=True)`. Create an issue in the PyTorch GitHub repository against the *onnx* component. Attach the pt2 model and the error report.
E           Error report has been saved to 'onnx_export_2024-06-21_21-40-55-111073_conversion.md'.
```

Refactor the export api.