Throw only torch.export errors (#81)
We throw e_export and not e_trace because we want to
focus on the torch.export.export error. The torch.jit.trace
error is due to the fallback and can be confusing to users.
e_trace is still saved in the error report and will include
e_export as well.