Update value names (#107)
Instead of using the output names from node.target._schema,
we always use the index if there are more than one outputs so the
names can be programmatically reconstructed. This is useful for
comparing values from the ONNX graph with those from the FX graph.

When there are multiple outputs, the output names will be
node_name__0, node_name__1, etc.